Complete exercise 4.21 - Add token auth when deleting blog posts

+++ part4/blog-list/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require("../models/blog");
 const User = require("../models/user");
 const jwt = require("jsonwebtoken");
 
-const getTokenFrom = (request) => {
-  const authorization = request.get("authorization");
-  if (authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "");
-  }
-  return null;
-};
-
 blogsRouter.get("/", async (request, response, next) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   return response.json(blogs);
@@ -52,9 +44,21 @@ blogsRouter.put("/:id", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
   const blog = await Blog.findById({ _id: request.params.id });
-  const result = await Blog.findByIdAndDelete(request.params.id);
-  return response.status(204).json(blog);
+
+  if (blog.user.toString() === user.id.toString()) {
+    const result = await Blog.findByIdAndDelete(request.params.id);
+    return response.status(204).json(result);
+  }
+  return response
+    .status(401)
+    .json({ error: "you can only delete your own blog posts" });
 });
 
 module.exports = blogsRouter;
